Refactor helper methods into rollover helper
These are not really specific to the provider
