+++ app/helpers/provider_helper.rb
@@ -21,18 +21,6 @@ module ProviderHelper
     end
   end
 
-  def rollover_inactive
-    !FeatureService.enabled?("rollover.can_edit_current_and_next_cycles")
-  end
-
-  def rollover_active_and_current_cycle?(recruitment_cycle_year)
-    FeatureService.enabled?("rollover.can_edit_current_and_next_cycles") && (recruitment_cycle_year.to_i == Settings.current_recruitment_cycle_year)
-  end
-
-  def rollover_active_and_next_cycle?(recruitment_cycle_year)
-    FeatureService.enabled?("rollover.can_edit_current_and_next_cycles") && (recruitment_cycle_year.to_i == Settings.current_recruitment_cycle_year + 1)
-  end
-
 private
 
   def visa_sponsorship_call_to_action(provider)
